% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{ads.list}
\alias{ads.list}
\title{Retrieves a list of ads, possibly filtered. This method supports paging.}
\usage{
ads.list(profileId, active = NULL, advertiserId = NULL, archived = NULL,
  audienceSegmentIds = NULL, campaignIds = NULL, compatibility = NULL,
  creativeIds = NULL, creativeOptimizationConfigurationIds = NULL,
  creativeType = NULL, dynamicClickTracker = NULL, ids = NULL,
  landingPageIds = NULL, maxResults = NULL, overriddenEventTagId = NULL,
  pageToken = NULL, placementIds = NULL, remarketingListIds = NULL,
  searchString = NULL, sizeIds = NULL, sortField = NULL,
  sortOrder = NULL, sslCompliant = NULL, sslRequired = NULL,
  type = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{active}{Select only active ads}

\item{advertiserId}{Select only ads with this advertiser ID}

\item{archived}{Select only archived ads}

\item{audienceSegmentIds}{Select only ads with these audience segment IDs}

\item{campaignIds}{Select only ads with these campaign IDs}

\item{compatibility}{Select default ads with the specified compatibility}

\item{creativeIds}{Select only ads with these creative IDs assigned}

\item{creativeOptimizationConfigurationIds}{Select only ads with these creative optimization configuration IDs}

\item{creativeType}{Select only ads with the specified creativeType}

\item{dynamicClickTracker}{Select only dynamic click trackers}

\item{ids}{Select only ads with these IDs}

\item{landingPageIds}{Select only ads with these landing page IDs}

\item{maxResults}{Maximum number of results to return}

\item{overriddenEventTagId}{Select only ads with this event tag override ID}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{placementIds}{Select only ads with these placement IDs assigned}

\item{remarketingListIds}{Select only ads whose list targeting expression use these remarketing list IDs}

\item{searchString}{Allows searching for objects by name or ID}

\item{sizeIds}{Select only ads with these size IDs}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{sslCompliant}{Select only ads that are SSL-compliant}

\item{sslRequired}{Select only ads that require SSL}

\item{type}{Select only ads with these types}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}
