% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AudienceSegment}
\alias{AudienceSegment}
\title{AudienceSegment Object}
\usage{
AudienceSegment(allocation = NULL, id = NULL, name = NULL)
}
\arguments{
\item{allocation}{Weight allocated to this segment}

\item{id}{ID of this audience segment}

\item{name}{Name of this audience segment}
}
\value{
AudienceSegment object
}
\description{
AudienceSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Audience Segment.
}

