% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{campaignCreativeAssociations.insert}
\alias{campaignCreativeAssociations.insert}
\title{Associates a creative with the specified campaign. This method creates a default ad with dimensions matching the creative in the campaign if such a default ad does not exist already.}
\usage{
campaignCreativeAssociations.insert(CampaignCreativeAssociation, profileId,
  campaignId)
}
\arguments{
\item{CampaignCreativeAssociation}{The \link{CampaignCreativeAssociation} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{campaignId}{Campaign ID in this association}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other CampaignCreativeAssociation functions: \code{\link{CampaignCreativeAssociation}}
}

