% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{campaigns.patch}
\alias{campaigns.patch}
\title{Updates an existing campaign. This method supports patch semantics.}
\usage{
campaigns.patch(Campaign, profileId, id)
}
\arguments{
\item{Campaign}{The \link{Campaign} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{Campaign ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other Campaign functions: \code{\link{Campaign}},
  \code{\link{campaigns.insert}},
  \code{\link{campaigns.update}}
}

