% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CampaignsListResponse}
\alias{CampaignsListResponse}
\title{CampaignsListResponse Object}
\usage{
CampaignsListResponse(campaigns = NULL, nextPageToken = NULL)
}
\arguments{
\item{campaigns}{Campaign collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
CampaignsListResponse object
}
\description{
CampaignsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Campaign List Response
}

