% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConversionsBatchInsertRequest}
\alias{ConversionsBatchInsertRequest}
\title{ConversionsBatchInsertRequest Object}
\usage{
ConversionsBatchInsertRequest(conversions = NULL, encryptionInfo = NULL)
}
\arguments{
\item{conversions}{The set of conversions to insert}

\item{encryptionInfo}{Describes how encryptedUserId or encryptedUserIdCandidates[] is encrypted}
}
\value{
ConversionsBatchInsertRequest object
}
\description{
ConversionsBatchInsertRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Insert Conversions Request.
}
\seealso{
Other ConversionsBatchInsertRequest functions: \code{\link{conversions.batchinsert}}
}

