% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeAsset}
\alias{CreativeAsset}
\title{CreativeAsset Object}
\usage{
CreativeAsset(actionScript3 = NULL, active = NULL, alignment = NULL,
  artworkType = NULL, assetIdentifier = NULL, backupImageExit = NULL,
  bitRate = NULL, childAssetType = NULL, collapsedSize = NULL,
  companionCreativeIds = NULL, customStartTimeValue = NULL,
  detectedFeatures = NULL, displayType = NULL, duration = NULL,
  durationType = NULL, expandedDimension = NULL, fileSize = NULL,
  flashVersion = NULL, hideFlashObjects = NULL, hideSelectionBoxes = NULL,
  horizontallyLocked = NULL, id = NULL, idDimensionValue = NULL,
  mimeType = NULL, offset = NULL, originalBackup = NULL,
  position = NULL, positionLeftUnit = NULL, positionTopUnit = NULL,
  progressiveServingUrl = NULL, pushdown = NULL, pushdownDuration = NULL,
  role = NULL, size = NULL, sslCompliant = NULL, startTimeType = NULL,
  streamingServingUrl = NULL, transparency = NULL,
  verticallyLocked = NULL, videoDuration = NULL, windowMode = NULL,
  zIndex = NULL, zipFilename = NULL, zipFilesize = NULL)
}
\arguments{
\item{actionScript3}{Whether ActionScript3 is enabled for the flash asset}

\item{active}{Whether the video asset is active}

\item{alignment}{Possible alignments for an asset}

\item{artworkType}{Artwork type of rich media creative}

\item{assetIdentifier}{Identifier of this asset}

\item{backupImageExit}{Exit event configured for the backup image}

\item{bitRate}{Detected bit-rate for video asset}

\item{childAssetType}{Rich media child asset type}

\item{collapsedSize}{Size of an asset when collapsed}

\item{companionCreativeIds}{List of companion creatives assigned to an in-stream video creative asset}

\item{customStartTimeValue}{Custom start time in seconds for making the asset visible}

\item{detectedFeatures}{List of feature dependencies for the creative asset that are detected by DCM}

\item{displayType}{Type of rich media asset}

\item{duration}{Duration in seconds for which an asset will be displayed}

\item{durationType}{Duration type for which an asset will be displayed}

\item{expandedDimension}{Detected expanded dimension for video asset}

\item{fileSize}{File size associated with this creative asset}

\item{flashVersion}{Flash version of the asset}

\item{hideFlashObjects}{Whether to hide Flash objects flag for an asset}

\item{hideSelectionBoxes}{Whether to hide selection boxes flag for an asset}

\item{horizontallyLocked}{Whether the asset is horizontally locked}

\item{id}{Numeric ID of this creative asset}

\item{idDimensionValue}{Dimension value for the ID of the asset}

\item{mimeType}{Detected MIME type for video asset}

\item{offset}{Offset position for an asset in collapsed mode}

\item{originalBackup}{Whether the backup asset is original or changed by the user in DCM}

\item{position}{Offset position for an asset}

\item{positionLeftUnit}{Offset left unit for an asset}

\item{positionTopUnit}{Offset top unit for an asset}

\item{progressiveServingUrl}{Progressive URL for video asset}

\item{pushdown}{Whether the asset pushes down other content}

\item{pushdownDuration}{Pushdown duration in seconds for an asset}

\item{role}{Role of the asset in relation to creative}

\item{size}{Size associated with this creative asset}

\item{sslCompliant}{Whether the asset is SSL-compliant}

\item{startTimeType}{Initial wait time type before making the asset visible}

\item{streamingServingUrl}{Streaming URL for video asset}

\item{transparency}{Whether the asset is transparent}

\item{verticallyLocked}{Whether the asset is vertically locked}

\item{videoDuration}{Detected video duration for video asset}

\item{windowMode}{Window mode options for flash assets}

\item{zIndex}{zIndex value of an asset}

\item{zipFilename}{File name of zip file}

\item{zipFilesize}{Size of zip file}
}
\value{
CreativeAsset object
}
\description{
CreativeAsset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Asset.
}

