% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeFieldValue}
\alias{CreativeFieldValue}
\title{CreativeFieldValue Object}
\usage{
CreativeFieldValue(id = NULL, value = NULL)
}
\arguments{
\item{id}{ID of this creative field value}

\item{value}{Value of this creative field value}
}
\value{
CreativeFieldValue object
}
\description{
CreativeFieldValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a creative field value.
}
\seealso{
Other CreativeFieldValue functions: \code{\link{creativeFieldValues.insert}},
  \code{\link{creativeFieldValues.patch}},
  \code{\link{creativeFieldValues.update}}
}

