% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeGroupAssignment}
\alias{CreativeGroupAssignment}
\title{CreativeGroupAssignment Object}
\usage{
CreativeGroupAssignment(creativeGroupId = NULL, creativeGroupNumber = NULL)
}
\arguments{
\item{creativeGroupId}{ID of the creative group to be assigned}

\item{creativeGroupNumber}{Creative group number of the creative group assignment}
}
\value{
CreativeGroupAssignment object
}
\description{
CreativeGroupAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Group Assignment.
}

