% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySiteContact}
\alias{DirectorySiteContact}
\title{DirectorySiteContact Object}
\usage{
DirectorySiteContact(address = NULL, email = NULL, firstName = NULL,
  id = NULL, lastName = NULL, phone = NULL, role = NULL, title = NULL,
  type = NULL)
}
\arguments{
\item{address}{Address of this directory site contact}

\item{email}{Email address of this directory site contact}

\item{firstName}{First name of this directory site contact}

\item{id}{ID of this directory site contact}

\item{lastName}{Last name of this directory site contact}

\item{phone}{Phone number of this directory site contact}

\item{role}{Directory site contact role}

\item{title}{Title or designation of this directory site contact}

\item{type}{Directory site contact type}
}
\value{
DirectorySiteContact object
}
\description{
DirectorySiteContact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a Site Directory contact.
}

