% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{dynamicTargetingKeys.list}
\alias{dynamicTargetingKeys.list}
\title{Retrieves a list of dynamic targeting keys.}
\usage{
dynamicTargetingKeys.list(profileId, advertiserId = NULL, names = NULL,
  objectId = NULL, objectType = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserId}{Select only dynamic targeting keys whose object has this advertiser ID}

\item{names}{Select only dynamic targeting keys exactly matching these names}

\item{objectId}{Select only dynamic targeting keys with this object ID}

\item{objectType}{Select only dynamic targeting keys with this object type}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

