% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DynamicTargetingKeysListResponse}
\alias{DynamicTargetingKeysListResponse}
\title{DynamicTargetingKeysListResponse Object}
\usage{
DynamicTargetingKeysListResponse(dynamicTargetingKeys = NULL)
}
\arguments{
\item{dynamicTargetingKeys}{Dynamic targeting key collection}
}
\value{
DynamicTargetingKeysListResponse object
}
\description{
DynamicTargetingKeysListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dynamic Targeting Key List Response
}

