% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{files.list}
\alias{files.list}
\title{Lists files for a user profile.}
\usage{
files.list(profileId, maxResults = NULL, pageToken = NULL, scope = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{The DFA profile ID}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{The value of the nextToken from the previous result page}

\item{scope}{The scope that defines which results are returned, default is 'MINE'}

\item{sortField}{The field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is 'DESCENDING'}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfareporting)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

