% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{floodlightActivities.generatetag}
\alias{floodlightActivities.generatetag}
\title{Generates a tag for a floodlight activity.}
\usage{
floodlightActivities.generatetag(profileId, floodlightActivityId = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{floodlightActivityId}{Floodlight activity ID for which we want to generate a tag}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

