% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivity}
\alias{FloodlightActivity}
\title{FloodlightActivity Object}
\usage{
FloodlightActivity(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, cacheBustingType = NULL,
  countingMethod = NULL, defaultTags = NULL, expectedUrl = NULL,
  floodlightActivityGroupId = NULL, floodlightActivityGroupName = NULL,
  floodlightActivityGroupTagString = NULL,
  floodlightActivityGroupType = NULL, floodlightConfigurationId = NULL,
  floodlightConfigurationIdDimensionValue = NULL, hidden = NULL,
  id = NULL, idDimensionValue = NULL, imageTagEnabled = NULL,
  name = NULL, notes = NULL, publisherTags = NULL, secure = NULL,
  sslCompliant = NULL, sslRequired = NULL, subaccountId = NULL,
  tagFormat = NULL, tagString = NULL, userDefinedVariableTypes = NULL)
}
\arguments{
\item{accountId}{Account ID of this floodlight activity}

\item{advertiserId}{Advertiser ID of this floodlight activity}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{cacheBustingType}{Code type used for cache busting in the generated tag}

\item{countingMethod}{Counting method for conversions for this floodlight activity}

\item{defaultTags}{Dynamic floodlight tags}

\item{expectedUrl}{URL where this tag will be deployed}

\item{floodlightActivityGroupId}{Floodlight activity group ID of this floodlight activity}

\item{floodlightActivityGroupName}{Name of the associated floodlight activity group}

\item{floodlightActivityGroupTagString}{Tag string of the associated floodlight activity group}

\item{floodlightActivityGroupType}{Type of the associated floodlight activity group}

\item{floodlightConfigurationId}{Floodlight configuration ID of this floodlight activity}

\item{floodlightConfigurationIdDimensionValue}{Dimension value for the ID of the floodlight configuration}

\item{hidden}{Whether this activity is archived}

\item{id}{ID of this floodlight activity}

\item{idDimensionValue}{Dimension value for the ID of this floodlight activity}

\item{imageTagEnabled}{Whether the image tag is enabled for this activity}

\item{name}{Name of this floodlight activity}

\item{notes}{General notes or implementation instructions for the tag}

\item{publisherTags}{Publisher dynamic floodlight tags}

\item{secure}{Whether this tag should use SSL}

\item{sslCompliant}{Whether the floodlight activity is SSL-compliant}

\item{sslRequired}{Whether this floodlight activity must be SSL-compliant}

\item{subaccountId}{Subaccount ID of this floodlight activity}

\item{tagFormat}{Tag format type for the floodlight activity}

\item{tagString}{Value of the cat= paramter in the floodlight tag, which the ad servers use to identify the activity}

\item{userDefinedVariableTypes}{List of the user-defined variables used by this conversion tag}
}
\value{
FloodlightActivity object
}
\description{
FloodlightActivity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a Floodlight activity.
}
\seealso{
Other FloodlightActivity functions: \code{\link{floodlightActivities.insert}},
  \code{\link{floodlightActivities.patch}},
  \code{\link{floodlightActivities.update}}
}

