% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivityDynamicTag}
\alias{FloodlightActivityDynamicTag}
\title{FloodlightActivityDynamicTag Object}
\usage{
FloodlightActivityDynamicTag(id = NULL, name = NULL, tag = NULL)
}
\arguments{
\item{id}{ID of this dynamic tag}

\item{name}{Name of this tag}

\item{tag}{Tag code}
}
\value{
FloodlightActivityDynamicTag object
}
\description{
FloodlightActivityDynamicTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dynamic Tag
}

