% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivityPublisherDynamicTag}
\alias{FloodlightActivityPublisherDynamicTag}
\title{FloodlightActivityPublisherDynamicTag Object}
\usage{
FloodlightActivityPublisherDynamicTag(clickThrough = NULL,
  directorySiteId = NULL, dynamicTag = NULL, siteId = NULL,
  siteIdDimensionValue = NULL, viewThrough = NULL)
}
\arguments{
\item{clickThrough}{Whether this tag is applicable only for click-throughs}

\item{directorySiteId}{Directory site ID of this dynamic tag}

\item{dynamicTag}{Dynamic floodlight tag}

\item{siteId}{Site ID of this dynamic tag}

\item{siteIdDimensionValue}{Dimension value for the ID of the site}

\item{viewThrough}{Whether this tag is applicable only for view-throughs}
}
\value{
FloodlightActivityPublisherDynamicTag object
}
\description{
FloodlightActivityPublisherDynamicTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Publisher Dynamic Tag
}

