% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{landingPages.patch}
\alias{landingPages.patch}
\title{Updates an existing campaign landing page. This method supports patch semantics.}
\usage{
landingPages.patch(LandingPage, profileId, campaignId, id)
}
\arguments{
\item{LandingPage}{The \link{LandingPage} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{campaignId}{Landing page campaign ID}

\item{id}{Landing page ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other LandingPage functions: \code{\link{LandingPage}},
  \code{\link{landingPages.insert}},
  \code{\link{landingPages.update}}
}

