% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Language}
\alias{Language}
\title{Language Object}
\usage{
Language(id = NULL, languageCode = NULL, name = NULL)
}
\arguments{
\item{id}{Language ID of this language}

\item{languageCode}{Format of language code is an ISO 639 two-letter language code optionally followed by an underscore followed by an ISO 3166 code}

\item{name}{Name of this language}
}
\value{
Language object
}
\description{
Language Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a language that can be targeted by ads.
}

