% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OffsetPosition}
\alias{OffsetPosition}
\title{OffsetPosition Object}
\usage{
OffsetPosition(left = NULL, top = NULL)
}
\arguments{
\item{left}{Offset distance from left side of an asset or a window}

\item{top}{Offset distance from top side of an asset or a window}
}
\value{
OffsetPosition object
}
\description{
OffsetPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Offset Position.
}

