% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OperatingSystem}
\alias{OperatingSystem}
\title{OperatingSystem Object}
\usage{
OperatingSystem(dartId = NULL, desktop = NULL, mobile = NULL,
  name = NULL)
}
\arguments{
\item{dartId}{DART ID of this operating system}

\item{desktop}{Whether this operating system is for desktop}

\item{mobile}{Whether this operating system is for mobile}

\item{name}{Name of this operating system}
}
\value{
OperatingSystem object
}
\description{
OperatingSystem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about an operating system that can be targeted by ads.
}

