% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Order}
\alias{Order}
\title{Order Object}
\usage{
Order(accountId = NULL, advertiserId = NULL,
  approverUserProfileIds = NULL, buyerInvoiceId = NULL,
  buyerOrganizationName = NULL, comments = NULL, contacts = NULL,
  id = NULL, lastModifiedInfo = NULL, name = NULL, notes = NULL,
  planningTermId = NULL, projectId = NULL, sellerOrderId = NULL,
  sellerOrganizationName = NULL, siteId = NULL, siteNames = NULL,
  subaccountId = NULL, termsAndConditions = NULL)
}
\arguments{
\item{accountId}{Account ID of this order}

\item{advertiserId}{Advertiser ID of this order}

\item{approverUserProfileIds}{IDs for users that have to approve documents created for this order}

\item{buyerInvoiceId}{Buyer invoice ID associated with this order}

\item{buyerOrganizationName}{Name of the buyer organization}

\item{comments}{Comments in this order}

\item{contacts}{Contacts for this order}

\item{id}{ID of this order}

\item{lastModifiedInfo}{Information about the most recent modification of this order}

\item{name}{Name of this order}

\item{notes}{Notes of this order}

\item{planningTermId}{ID of the terms and conditions template used in this order}

\item{projectId}{Project ID of this order}

\item{sellerOrderId}{Seller order ID associated with this order}

\item{sellerOrganizationName}{Name of the seller organization}

\item{siteId}{Site IDs this order is associated with}

\item{siteNames}{Free-form site names this order is associated with}

\item{subaccountId}{Subaccount ID of this order}

\item{termsAndConditions}{Terms and conditions of this order}
}
\value{
Order object
}
\description{
Order Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes properties of a DoubleClick Planning order.
}

