% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{placements.list}
\alias{placements.list}
\title{Retrieves a list of placements, possibly filtered.}
\usage{
placements.list(profileId, advertiserIds = NULL, archived = NULL,
  campaignIds = NULL, compatibilities = NULL, contentCategoryIds = NULL,
  directorySiteIds = NULL, groupIds = NULL, ids = NULL,
  maxEndDate = NULL, maxResults = NULL, maxStartDate = NULL,
  minEndDate = NULL, minStartDate = NULL, pageToken = NULL,
  paymentSource = NULL, placementStrategyIds = NULL, pricingTypes = NULL,
  searchString = NULL, siteIds = NULL, sizeIds = NULL, sortField = NULL,
  sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserIds}{Select only placements that belong to these advertisers}

\item{archived}{Select only archived placements}

\item{campaignIds}{Select only placements that belong to these campaigns}

\item{compatibilities}{Select only placements that are associated with these compatibilities}

\item{contentCategoryIds}{Select only placements that are associated with these content categories}

\item{directorySiteIds}{Select only placements that are associated with these directory sites}

\item{groupIds}{Select only placements that belong to these placement groups}

\item{ids}{Select only placements with these IDs}

\item{maxEndDate}{Select only placements or placement groups whose end date is on or before the specified maxEndDate}

\item{maxResults}{Maximum number of results to return}

\item{maxStartDate}{Select only placements or placement groups whose start date is on or before the specified maxStartDate}

\item{minEndDate}{Select only placements or placement groups whose end date is on or after the specified minEndDate}

\item{minStartDate}{Select only placements or placement groups whose start date is on or after the specified minStartDate}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{paymentSource}{Select only placements with this payment source}

\item{placementStrategyIds}{Select only placements that are associated with these placement strategies}

\item{pricingTypes}{Select only placements with these pricing types}

\item{searchString}{Allows searching for placements by name or ID}

\item{siteIds}{Select only placements that are associated with these sites}

\item{sizeIds}{Select only placements that are associated with these sizes}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

