% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementStrategy}
\alias{PlacementStrategy}
\title{PlacementStrategy Object}
\usage{
PlacementStrategy(accountId = NULL, id = NULL, name = NULL)
}
\arguments{
\item{accountId}{Account ID of this placement strategy}

\item{id}{ID of this placement strategy}

\item{name}{Name of this placement strategy}
}
\value{
PlacementStrategy object
}
\description{
PlacementStrategy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a placement strategy.
}
\seealso{
Other PlacementStrategy functions: \code{\link{placementStrategies.insert}},
  \code{\link{placementStrategies.patch}},
  \code{\link{placementStrategies.update}}
}

