% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PostalCode}
\alias{PostalCode}
\title{PostalCode Object}
\usage{
PostalCode(code = NULL, countryCode = NULL, countryDartId = NULL,
  id = NULL)
}
\arguments{
\item{code}{Postal code}

\item{countryCode}{Country code of the country to which this postal code belongs}

\item{countryDartId}{DART ID of the country to which this postal code belongs}

\item{id}{ID of this postal code}
}
\value{
PostalCode object
}
\description{
PostalCode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a postal code that can be targeted by ads.
}

