% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PricingSchedule}
\alias{PricingSchedule}
\title{PricingSchedule Object}
\usage{
PricingSchedule(capCostOption = NULL, disregardOverdelivery = NULL,
  endDate = NULL, flighted = NULL, floodlightActivityId = NULL,
  pricingPeriods = NULL, pricingType = NULL, startDate = NULL,
  testingStartDate = NULL)
}
\arguments{
\item{capCostOption}{Placement cap cost option}

\item{disregardOverdelivery}{Whether cap costs are ignored by ad serving}

\item{endDate}{Placement end date}

\item{flighted}{Whether this placement is flighted}

\item{floodlightActivityId}{Floodlight activity ID associated with this placement}

\item{pricingPeriods}{Pricing periods for this placement}

\item{pricingType}{Placement pricing type}

\item{startDate}{Placement start date}

\item{testingStartDate}{Testing start date of this placement}
}
\value{
PricingSchedule object
}
\description{
PricingSchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pricing Schedule
}

