% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Project}
\alias{Project}
\title{Project Object}
\usage{
Project(accountId = NULL, advertiserId = NULL, audienceAgeGroup = NULL,
  audienceGender = NULL, budget = NULL, clientBillingCode = NULL,
  clientName = NULL, endDate = NULL, id = NULL, lastModifiedInfo = NULL,
  name = NULL, overview = NULL, startDate = NULL, subaccountId = NULL,
  targetClicks = NULL, targetConversions = NULL, targetCpaNanos = NULL,
  targetCpcNanos = NULL, targetCpmActiveViewNanos = NULL,
  targetCpmNanos = NULL, targetImpressions = NULL)
}
\arguments{
\item{accountId}{Account ID of this project}

\item{advertiserId}{Advertiser ID of this project}

\item{audienceAgeGroup}{Audience age group of this project}

\item{audienceGender}{Audience gender of this project}

\item{budget}{Budget of this project in the currency specified by the current account}

\item{clientBillingCode}{Client billing code of this project}

\item{clientName}{Name of the project client}

\item{endDate}{End date of the project}

\item{id}{ID of this project}

\item{lastModifiedInfo}{Information about the most recent modification of this project}

\item{name}{Name of this project}

\item{overview}{Overview of this project}

\item{startDate}{Start date of the project}

\item{subaccountId}{Subaccount ID of this project}

\item{targetClicks}{Number of clicks that the advertiser is targeting}

\item{targetConversions}{Number of conversions that the advertiser is targeting}

\item{targetCpaNanos}{CPA that the advertiser is targeting}

\item{targetCpcNanos}{CPC that the advertiser is targeting}

\item{targetCpmActiveViewNanos}{vCPM from Active View that the advertiser is targeting}

\item{targetCpmNanos}{CPM that the advertiser is targeting}

\item{targetImpressions}{Number of impressions that the advertiser is targeting}
}
\value{
Project object
}
\description{
Project Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a DoubleClick Planning project.
}

