% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{reports.delete}
\alias{reports.delete}
\title{Deletes a report by its ID.}
\usage{
reports.delete(profileId, reportId)
}
\arguments{
\item{profileId}{The DFA user profile ID}

\item{reportId}{The ID of the report}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfareporting)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

