% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{reports.insert}
\alias{reports.insert}
\title{Creates a report.}
\usage{
reports.insert(Report, profileId)
}
\arguments{
\item{Report}{The \link{Report} object to pass to this method}

\item{profileId}{The DFA user profile ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfareporting)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}

Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}

