% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{reports.run}
\alias{reports.run}
\title{Runs a report.}
\usage{
reports.run(profileId, reportId, synchronous = NULL)
}
\arguments{
\item{profileId}{The DFA profile ID}

\item{reportId}{The ID of the report}

\item{synchronous}{If set and true, tries to run the report synchronously}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfareporting)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

