% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RichMediaExitOverride}
\alias{RichMediaExitOverride}
\title{RichMediaExitOverride Object}
\usage{
RichMediaExitOverride(clickThroughUrl = NULL, enabled = NULL,
  exitId = NULL)
}
\arguments{
\item{clickThroughUrl}{Click-through URL of this rich media exit override}

\item{enabled}{Whether to use the clickThroughUrl}

\item{exitId}{ID for the override to refer to a specific exit in the creative}
}
\value{
RichMediaExitOverride object
}
\description{
RichMediaExitOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rich Media Exit Override.
}

