% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Rule}
\alias{Rule}
\title{Rule Object}
\usage{
Rule(assetId = NULL, name = NULL, targetingTemplateId = NULL)
}
\arguments{
\item{assetId}{A creativeAssets[]}

\item{name}{A user-friendly name for this rule}

\item{targetingTemplateId}{A targeting template ID}
}
\value{
Rule object
}
\description{
Rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule associates an asset with a targeting template for asset-level targeting. Applicable to INSTREAM_VIDEO creatives.
}

