% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SiteContact}
\alias{SiteContact}
\title{SiteContact Object}
\usage{
SiteContact(address = NULL, contactType = NULL, email = NULL,
  firstName = NULL, id = NULL, lastName = NULL, phone = NULL,
  title = NULL)
}
\arguments{
\item{address}{Address of this site contact}

\item{contactType}{Site contact type}

\item{email}{Email address of this site contact}

\item{firstName}{First name of this site contact}

\item{id}{ID of this site contact}

\item{lastName}{Last name of this site contact}

\item{phone}{Primary phone number of this site contact}

\item{title}{Title or designation of this site contact}
}
\value{
SiteContact object
}
\description{
SiteContact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Site Contact
}

