% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{sites.list}
\alias{sites.list}
\title{Retrieves a list of sites, possibly filtered.}
\usage{
sites.list(profileId, acceptsInStreamVideoPlacements = NULL,
  acceptsInterstitialPlacements = NULL,
  acceptsPublisherPaidPlacements = NULL, adWordsSite = NULL,
  approved = NULL, campaignIds = NULL, directorySiteIds = NULL,
  ids = NULL, maxResults = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL, subaccountId = NULL,
  unmappedSite = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{acceptsInStreamVideoPlacements}{This search filter is no longer supported and will have no effect on the results returned}

\item{acceptsInterstitialPlacements}{This search filter is no longer supported and will have no effect on the results returned}

\item{acceptsPublisherPaidPlacements}{Select only sites that accept publisher paid placements}

\item{adWordsSite}{Select only AdWords sites}

\item{approved}{Select only approved sites}

\item{campaignIds}{Select only sites with these campaign IDs}

\item{directorySiteIds}{Select only sites with these directory site IDs}

\item{ids}{Select only sites with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name, ID or keyName}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{subaccountId}{Select only sites with this subaccount ID}

\item{unmappedSite}{Select only sites that have not been mapped to a directory site}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

