% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{sizes.list}
\alias{sizes.list}
\title{Retrieves a list of sizes, possibly filtered.}
\usage{
sizes.list(profileId, height = NULL, iabStandard = NULL, ids = NULL,
  width = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{height}{Select only sizes with this height}

\item{iabStandard}{Select only IAB standard sizes}

\item{ids}{Select only sizes with these IDs}

\item{width}{Select only sizes with this width}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

