% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TargetingTemplate}
\alias{TargetingTemplate}
\title{TargetingTemplate Object}
\usage{
TargetingTemplate(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, dayPartTargeting = NULL,
  geoTargeting = NULL, id = NULL, keyValueTargetingExpression = NULL,
  languageTargeting = NULL, listTargetingExpression = NULL, name = NULL,
  subaccountId = NULL, technologyTargeting = NULL)
}
\arguments{
\item{accountId}{Account ID of this targeting template}

\item{advertiserId}{Advertiser ID of this targeting template}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{dayPartTargeting}{Time and day targeting criteria}

\item{geoTargeting}{Geographical targeting criteria}

\item{id}{ID of this targeting template}

\item{keyValueTargetingExpression}{Key-value targeting criteria}

\item{languageTargeting}{Language targeting criteria}

\item{listTargetingExpression}{Remarketing list targeting criteria}

\item{name}{Name of this targeting template}

\item{subaccountId}{Subaccount ID of this targeting template}

\item{technologyTargeting}{Technology platform targeting criteria}
}
\value{
TargetingTemplate object
}
\description{
TargetingTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a targeting template. A targeting template encapsulates targeting information which can be reused across multiple ads.
}
\seealso{
Other TargetingTemplate functions: \code{\link{targetingTemplates.insert}},
  \code{\link{targetingTemplates.patch}},
  \code{\link{targetingTemplates.update}}
}

