% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TargetingTemplatesListResponse}
\alias{TargetingTemplatesListResponse}
\title{TargetingTemplatesListResponse Object}
\usage{
TargetingTemplatesListResponse(nextPageToken = NULL,
  targetingTemplates = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{targetingTemplates}{Targeting template collection}
}
\value{
TargetingTemplatesListResponse object
}
\description{
TargetingTemplatesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Targeting Template List Response
}

