% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TechnologyTargeting}
\alias{TechnologyTargeting}
\title{TechnologyTargeting Object}
\usage{
TechnologyTargeting(browsers = NULL, connectionTypes = NULL,
  mobileCarriers = NULL, operatingSystemVersions = NULL,
  operatingSystems = NULL, platformTypes = NULL)
}
\arguments{
\item{browsers}{Browsers that this ad targets}

\item{connectionTypes}{Connection types that this ad targets}

\item{mobileCarriers}{Mobile carriers that this ad targets}

\item{operatingSystemVersions}{Operating system versions that this ad targets}

\item{operatingSystems}{Operating systems that this ad targets}

\item{platformTypes}{Platform types that this ad targets}
}
\value{
TechnologyTargeting object
}
\description{
TechnologyTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Technology Targeting.
}

