% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{userRoles.list}
\alias{userRoles.list}
\title{Retrieves a list of user roles, possibly filtered.}
\usage{
userRoles.list(profileId, accountUserRoleOnly = NULL, ids = NULL,
  maxResults = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL, subaccountId = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{accountUserRoleOnly}{Select only account level user roles not associated with any specific subaccount}

\item{ids}{Select only user roles with the specified IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name or ID}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{subaccountId}{Select only user roles that belong to this subaccount}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/}{Google Documentation}
}

