% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRolesListResponse}
\alias{UserRolesListResponse}
\title{UserRolesListResponse Object}
\usage{
UserRolesListResponse(nextPageToken = NULL, userRoles = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{userRoles}{User role collection}
}
\value{
UserRolesListResponse object
}
\description{
UserRolesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User Role List Response
}

